backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.19
django-import-export==3.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
exceptiongroup==1.1.3
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
Pillow==10.0.0
pluggy==1.3.0
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tablib==3.5.0
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
xlrd==2.0.1
xlwt==1.3.0
